gtk: make *_set_sort_func use gtk::Ordering
In rust, all the cmp functions returns a std::cmp::Ordering
which we can convert to a rust gtk::Ordering. The C functions though
can return a random int > 0, < 0 or 0. Instead of accepting a random int
here, request a gtk::Ordering which makes more sense for rust bindings.
