+++ gtk4/src/tree_sortable.rs
@@ -1,8 +1,8 @@
 // Take a look at the license at the top of the repository in the LICENSE file.
 
+use crate::Ordering;
 use glib::object::{Cast, IsA};
 use glib::translate::*;
-use std::cmp::Ordering;
 use std::fmt;
 use std::mem;
